Split search combobox styles
